Add the <Case> class
Added the <Case> class that represents a case in a mathematical proof.
For example if we want to prove a statement alpha(n) and we have to
consider cases where n = 2k and n = 2k + 1 for some natural number k,
then we will have two <Case> instances, one of them will hold the
information that n = 2k, the other one that n = 2k + 1

WARNING: the class is not tested at all!