Compute the GDC when solving relations
modified the method <LinearRelation.solve> so that after solving the
relation it devides the formulas by the greatest common divisor of
their multipliers.

Modified the method <LinearFormula.status> to take into account that
during solving of the relation, the relation might reverse.

Added the <gdc> function to misc.py that returnd the GDC of more or
less than two numbers.