Add the <recursive> option to the <get_bounds> method
Modified the method <LinearFormula.get_bounds> so that if an argument
<recursive> is set to True it will generate the bounds of the formula
recursively until all passed bounds will be used. This mins that
<LinearFormula('a+b').get_bounds(
	upper_bounds={'a': 'b', 'b': 'c'}, recursive=True)>
will return 'c + c' as the upper bound instead of 'b + c'.

Alse the <upper_bounds> and <lower_bounds> arguments are now set by
default to an empty dict.