Add the <NTermRecursionSequence.get_edge> method
This method gives an index as an argument, and returns a formula that
would be assigned to an edge of a graph in a SSA numbering if the
index-th and (index + 1)-th formulas of the sequence would be assigned
to the edge's ends. Generally the returned formula is a sum of the
index-th and (index + 1)-th formulas.